+++ gems/aws-sdk-core/lib/aws-sdk-core/log/param_filter.rb
@@ -26,13 +26,21 @@ module Aws
 
       def filter(values, type)
         case values
-        when Struct, Hash then filter_hash(values, type)
+        when Struct then filter_struct(values, type)
+        when Hash then filter_hash(values, type)
         when Array then filter_array(values, type)
         else values
         end
       end
 
       private
+
+      def filter_struct(values, type)
+        if values.class.include? Aws::Structure::Union
+          values = { values.member => values.value }
+        end
+        filter_hash(values, type)
+      end
 
       def filter_hash(values, type)
         if type.const_defined?('SENSITIVE')